location-plugin: introduce seperate _private folder and deprecate public utility functions
