+++ addon/plugins/location-plugin/utils/node-utils.ts
@@ -1,19 +1,10 @@
-import { SayController, NodeSelection, PNode } from '@lblod/ember-rdfa-editor';
-import { sayDataFactory } from '@lblod/ember-rdfa-editor/core/say-data-factory';
-import {
-  PROV,
-  RDF,
-} from '@lblod/ember-rdfa-editor-lblod-plugins/utils/constants';
-import { v4 as uuidv4 } from 'uuid';
-import { findAncestors } from '@lblod/ember-rdfa-editor/utils/position-utils';
-import {
-  Resource,
-  hasOutgoingNamedNodeTriple,
-} from '@lblod/ember-rdfa-editor-lblod-plugins/utils/namespace';
+import { SayController } from '@lblod/ember-rdfa-editor';
+import { Resource } from '@lblod/ember-rdfa-editor-lblod-plugins/utils/namespace';
 import { Area, Place } from './geo-helpers';
 import { Address } from './address-helpers';
-
+import { replaceSelectionWithLocation as replaceSelectionWithLocationInternal } from '../_private/utils/node-utils';
 /**
+ * @deprecated
  * Creates an 'OSLO location' node in place of the selection, along with the RDFa to create a triple
  * with the nearest parent of one of the passed types as the subject and the predicate
  * prov:atLocation. This doesn't work well with the RDFa tools, but since refactoring is required to
@@ -27,56 +18,13 @@ import { Address } from './address-helpers';
  */
 export function replaceSelectionWithLocation(
   controller: SayController,
-  subject: string,
-  toInsert?: Place | Address | Area,
+  toInsert: Place | Address | Area,
   subjectTypes?: Resource[],
 ) {
-  let resourceToLink: { pos: number; node: PNode } | undefined;
-  subjectTypes?.forEach((subjectType) => {
-    if (!resourceToLink) {
-      resourceToLink = findAncestors(
-        controller.mainEditorState.selection.$from,
-        (node) => {
-          if ('typeof' in node.attrs) {
-            return subjectType.matches(node.attrs.typeof);
-          }
-          return hasOutgoingNamedNodeTriple(
-            node.attrs,
-            RDF('type'),
-            subjectType,
-          );
-        },
-      )[0];
-    }
-  });
-  const backlinks = !resourceToLink
-    ? []
-    : [
-        {
-          predicate: PROV('atLocation').full,
-          subject: sayDataFactory.resourceNode(
-            resourceToLink.node.attrs.subject,
-          ),
-        },
-      ];
-
-  controller.withTransaction((tr) => {
-    tr.replaceSelectionWith(
-      controller.schema.node('oslo_location', {
-        subject: subject,
-        rdfaNodeType: 'resource',
-        __rdfaId: uuidv4(),
-        value: toInsert,
-        properties: [],
-        backlinks,
-      }),
-    );
-    if (tr.selection.$anchor.nodeBefore) {
-      const resolvedPos = tr.doc.resolve(
-        tr.selection.anchor - tr.selection.$anchor.nodeBefore?.nodeSize,
-      );
-      tr.setSelection(new NodeSelection(resolvedPos));
-    }
-    return tr;
-  });
+  return replaceSelectionWithLocationInternal(
+    controller,
+    toInsert.uri,
+    toInsert,
+    subjectTypes,
+  );
 }
